requirements.txt
django
django-ckeditor
pyyaml
canvasapi
